HSEARCH-3123 Move backend field name utils to the util.impl packages
Otherwise using these utils from packages the model depends on
(lucene.types.codec.impl in particular) leads to a cyclic dependency.
