+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/search/query/impl/IndexSensitiveHitExtractor.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.lucene.document.Document;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFields;
+import org.hibernate.search.v6poc.backend.lucene.util.impl.LuceneFields;
 
 class IndexSensitiveHitExtractor<C> implements HitExtractor<C> {
