+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/codec/impl/GeoPointFieldCodec.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.codec.impl;
 
-import static org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFields.internalFieldName;
+import static org.hibernate.search.v6poc.backend.lucene.util.impl.LuceneFields.internalFieldName;
 
 import java.util.Collections;
 import java.util.Objects;
